添加了 native 模块,已完成单元测试.
现在 wasm 可以跟 rust 的函数互相调用了.

+++ crates/engine/src/ins_branch.rs
@@ -260,5 +260,5 @@ pub fn exec_return(vm_module: Rc<RefCell<VMModule>>) -> Result<(), EngineError> 
             .get_relative_depth()
     };
 
-    br(vm_module, relative_depth)
+    br(vm_module, relative_depth as u32)
 }
